% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{Epimodelling}
\alias{Epimodelling}
\title{Package for analysis, simulation, visualisation of Compartmental Epidemiological Models}
\description{
The package is currently under development
}
\section{simulate data under the model}{

\emph{solve_SEIR}
}

\section{visualise the SEIR results}{

\emph{plot_SEIR}
}

\section{interactive shiny application based on SEIR}{

\emph{interact_SEIR}
}

