% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodellR_OOP.R
\name{epimodellR}
\alias{epimodellR}
\title{Set up a Compartmental Epidemiological Model}
\usage{
epimodellR(type = c("SIR", "SIS", "SIRD", "SEIR", "SEIS"))
}
\arguments{
\item{type}{Type of the model. You can choose from c("SIR", "SIS", "SIRD", "SEIR", "SEIS")}
}
\value{
The function \emph{epimodellR()} return an \emph{Epimodel} object. First, you need to specify the variable and parameter values in the console!
The list contains the \emph{type}, the  \emph{variables}, the \emph{var_values}, the \emph{parameters}, the \emph{par_values} and the system of differential \emph{equations} of the given model.
}
\description{
The function \emph{epimodellR()} helps you set up your compartmental epidemiological model
}
\examples{

mod <- epimodellR("SEIR")
 #parameter values
mod$par_values

}
