% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OOP.R
\name{plot.Epimodelled}
\alias{plot.Epimodelled}
\title{Method for plotting Epimodels}
\usage{
\method{plot}{Epimodelled}(res, main = "")
}
\arguments{
\item{res}{The results of the \emph{solve()} method, which is an \emph{Epimodelled} object}

\item{main}{Title for the plot}
}
\description{
The method plots the results of the solved system of differential equations based on the \emph{Epimodel} object.
}
\examples{

mod <- epimodellR("SIR")
mod_solved <- solve(mod)
plot(mod_solved)

}
