% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveR_OOP.R
\name{solve.Epimodel}
\alias{solve.Epimodel}
\title{Method for solving Epimodels}
\usage{
\method{solve}{Epimodel}(obj, delta_t = 0.01, days = 10)
}
\arguments{
\item{obj}{Epimodel object from the function \emph{epimodellR()}}

\item{delta_t}{rate of change in time (default = 0.01)}

\item{days}{number of days to simulate the model}
}
\value{
The function returns an \emph{Epimodelled} object which is a matrix containing the variables and the population size in every timestep
}
\description{
The solve is a method to \emph{solve()} the system of differential equations based on the \emph{Epimodel} object which encapsulates
the starting values, parameters and the system of differential equations.
}
\examples{
mod <- epimodellR("SIR")
mod_solved <- solve(mod)
head(mod_solved)


}
