% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_SEIR.R
\name{solve_SEIR}
\alias{solve_SEIR}
\title{Solve the SEIR system of differential equations}
\usage{
solve_SEIR(N, S0, E0, I0, mu, beta, a, gamma, delta_t = 0.01, times = 1000)
}
\arguments{
\item{N}{Population size}

\item{S0}{Initial number of the susceptible people}

\item{E0}{Initial number of the Exposed people}

\item{I0}{Initial number of Infectious people (N = S0 + E0 + I0 + R0 is assumed)}

\item{mu}{Mortality rate (for simplicity birth rate and mortality is
assumed to be equal)}

\item{beta}{Contact rate}

\item{a}{Incubation period}

\item{gamma}{Recovery rate}

\item{delta_t}{The rate of change in the time.}

\item{times}{The number of iteration (delta_t * times = days)}
}
\value{
A \emph{times} length list of the values of the model parameters:
S, E, I, R and delta_t. This list is the numerical solution of the system.
}
\description{
The function \emph{solve_SEIR} simulates data
under the SEIR model
}
\examples{
res <- solve_SEIR(1000, 500, 300, 200, 0.05, 0.8, 1, 0.3)
head(res)

}
