% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{Epimodelling}
\alias{Epimodelling}
\title{Package for simulation and visualisation of Compartmental Epidemiological Models}
\description{
The package has two main functionality: manual and interactive.
}
\section{Set up the model}{

\emph{epimodellR()}
}

\section{Solve the system of differential equations of the model}{

\emph{solve.Epimodel()}
}

\section{Visualise the solved model}{

\emph{plot.Epimodelled()}
}

\section{Interactive shiny application of the compartmental epidemiological models}{

\emph{EpiSimulator}
}

